Fix compiling errors for windows(mingw)

+++ lib/libgtkwave/src/gw-fst-loader.c
@@ -753,7 +753,7 @@ static GwDumpFile *gw_fst_loader_load(GwLoader *loader, const char *fname, GErro
     self->subvar_jrb = make_jrb(); /* only used for attributes such as generated in VHDL, etc. */
     self->synclock_jrb = make_jrb(); /* only used for synthetic clocks */
 
-    uint numfacs = fstReaderGetVarCount(self->fst_reader);
+    uint64_t numfacs = fstReaderGetVarCount(self->fst_reader);
 
     GwFacs *facs = gw_facs_new(numfacs);
     self->mvlfacs = g_new0(GwFac, numfacs);
@@ -761,7 +761,7 @@ static GwDumpFile *gw_fst_loader_load(GwLoader *loader, const char *fname, GErro
     node_block = g_new0(GwNode, numfacs);
     self->mvlfacs_rvs_alias = g_new0(fstHandle, numfacs);
 
-    fprintf(stderr, FST_RDLOAD "Processing %d facs.\n", numfacs);
+    fprintf(stderr, FST_RDLOAD "Processing %lu facs.\n", numfacs);
     // TODO: update splash
     // /* SPLASH */ splash_sync(1, 5);
